Support multiple format in internal catalog
